+++ tests/testthat/test_getBreederPed.R
@@ -1,9 +1,9 @@
 #' Copyright(c) 2017-2019 R. Mark Sharp
 #' This file is part of nprcmanager
-context("getBreederPed")
+context("getFocalAnimalPed")
 
-test_that("getBreederPed recognizes no file and wrong file arguments", {
-  expect_error(getBreederPed(), "\"fileName\" is missing, with no default")
-  expect_error(suppressWarnings(getBreederPed(fileName = "breeding file.csv")),
+test_that("getFocalAnimalPed recognizes no file and wrong file arguments", {
+  expect_error(getFocalAnimalPed(), "\"fileName\" is missing, with no default")
+  expect_error(suppressWarnings(getFocalAnimalPed(fileName = "breeding file.csv")),
                "cannot open the connection")
 })